Modified and added Images that will be used in patient login page and registration page, and renaming file name and relocating them to a specific directory for an easy configurations.
