--- admin-system/src/main/java/com/capstone/dentalclinic/demo/controller/PatientController/webPages.java
@@ -1,20 +1,0 @@
-package com.capstone.dentalclinic.demo.controller.PatientController;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import lombok.AllArgsConstructor;
-
-@Controller
-@RequestMapping
-@AllArgsConstructor
-public class webPages {
-    
-    @GetMapping("/")
-    public String landingpage() {
-        return "PatientWebPages/index";
-    }
-
-
-}
