+++ admin-system/src/main/java/com/capstone/dentalclinic/demo/services/administrator/ConfirmationTokenService.java
@@ -1,4 +1,4 @@
-package com.capstone.dentalclinic.demo.services;
+package com.capstone.dentalclinic.demo.services.administrator;
 
 import com.capstone.dentalclinic.demo.model.admin.token.ConfirmationToken;
