+++ admin-system/src/main/java/com/capstone/dentalclinic/demo/services/administrator/EmployeeServiceImpl.java
@@ -1,4 +1,4 @@
-package com.capstone.dentalclinic.demo.services;
+package com.capstone.dentalclinic.demo.services.administrator;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -19,7 +19,7 @@ import com.capstone.dentalclinic.demo.model.Roles;
 import com.capstone.dentalclinic.demo.model.admin.token.ConfirmationToken;
 import com.capstone.dentalclinic.demo.repository.EmployeeRepository;
 import com.capstone.dentalclinic.demo.security.PasswordEncoder;
-import com.capstone.dentalclinic.demo.services.email_template.EmailTemplate;
+import com.capstone.dentalclinic.demo.mail.email_template.EmailTemplate;
 
 import lombok.AllArgsConstructor;
