+++ admin-system/src/main/resources/static/assets/administrator/approved.png
