+++ admin-system/src/main/resources/static/assets/administrator/cancelled.png
