+++ admin-system/src/main/resources/static/assets/administrator/checklist.png
