+++ admin-system/src/main/resources/static/assets/administrator/dashboard.png
