+++ admin-system/src/main/resources/static/assets/administrator/new-resized-logo.png
