+++ admin-system/src/main/resources/static/assets/patients/PatientRegistrationPage.css
@@ -1,0 +1,120 @@
+@import url('https://fonts.googleapis.com/css2?family=Nunito:wght@400;500;600;700;800&display=swap');
+
+:root {
+    --primary: #fcd2d9;
+    --secondary: #faa4b1;
+    --mandyspink: #fb7c9a;
+    --lightprink: #f9dee3;
+    --white: #fffdfa;
+    --black: #070606;
+}
+
+html {
+    height: 100vh;
+    min-height: 100%;
+}
+
+body {
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+}
+
+/* 
+==================================================================
+                            Typography
+==================================================================
+*/
+
+* {
+    font-family: 'Nunito', sans-serif;
+    padding: 0;
+    margin: 0;
+}
+
+::placeholder {
+    font-family: 'Nunito', sans-serif;
+    font-weight: 500;
+    font-size: 1rem;
+}
+
+/* 
+==================================================================
+                            Container
+==================================================================
+*/
+
+.container {
+    max-width: 1500px !important;
+    width: 95% !important;
+    margin: 0 auto !important;
+}
+
+.form-container {
+    max-width: 1500px ;
+    width: 95%;
+    margin: 5rem auto 0 auto;
+    padding: 0 0 5rem 0;
+}
+
+/* 
+==================================================================
+                            Back Button
+==================================================================
+*/
+
+.back-button {
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    margin: 2rem 0 0 0;
+}
+
+.back-button img {
+    width: 40px;
+    image-rendering: optimizeQuality;
+}
+
+.back-button a{
+    text-transform: uppercase;
+    text-decoration: none;
+    font-weight: 500;
+    font-size: .800rem;
+    color: var(--black);
+    margin: 0 0 0 -.4rem;
+}
+
+.form-container-heading h1{
+    font-weight: 800;
+    font-size: 3rem;
+    margin: 0 0 0 0;
+}
+
+.form-container-heading p {
+    margin: 0 0 2rem 0;
+}
+
+.btn {
+    --bs-btn-padding-x: 1.4rem !important;
+    --bs-btn-padding-y: .400rem !important;
+    --bs-btn-font-family: 'Nunito' sans-serif !important;
+    --bs-btn-font-size: 1.1rem !important;
+    --bs-btn-font-weight: 500 !important;
+}
+
+.btn-outline-primary {
+    --bs-btn-color: var(--mandyspink) !important;
+    --bs-btn-border-color: var(--mandyspink) !important;
+    --bs-btn-hover-color: #fff;
+    --bs-btn-hover-bg: var(--mandyspink) !important;
+    --bs-btn-hover-border-color: var(--primary) !important;
+    --bs-btn-focus-shadow-rgb: 13,110,253;
+    --bs-btn-active-color: #fff;
+    --bs-btn-active-bg: var(--mandyspink) !important;
+    --bs-btn-active-border-color: var(--primary) !important;
+    --bs-btn-active-shadow: inset 0 3px 5pxrgba(0, 0, 0, 0.125);
+    /* --bs-btn-disabled-color: #0d6efd;
+    --bs-btn-disabled-bg: transparent;
+    --bs-btn-disabled-border-color: #0d6efd;
+    --bs-gradient: none; */
+}
