+++ admin-system/src/main/resources/static/javascript/administrator/dashboard.js
