+++ admin-system/src/main/resources/templates/PatientWebPages/PatientRegistrationPage.html
@@ -6,8 +6,15 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Gillego-Flores Dental Clinic</title>
 
+    <link rel="icon" type="image/png" sizes="32x32" href="/assets/administrator/new-resized-logo.png">
+    <link rel="shortcut icon" type="image/x-icon" sizes="16x16" href="/assets/administrator/new-resized-logo.png">
+    <link rel="manifest" href="/site.webmanifest">
+    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5">
+    <meta name="msapplication-TileColor" content="#da532c">
+    <meta name="theme-color" content="#ffffff">
+
     <!-- custom css-->
-    <link rel="stylesheet" href="css/PatientRegistrationPage.css">
+    <link rel="stylesheet" href="css/patients/PatientRegistrationPage.css">
 
     <!-- Bootstrap cdn-->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
@@ -36,10 +43,10 @@
 
         <div class="container">
             <div class="back-button">
-                <a href="#">
-                    <img src="img/back-arrow.svg" alt="back-arrow" />
+                <a href="/login">
+                    <img src="/assets/patients/back-arrow.svg" alt="back-arrow" />
                 </a>
-                <a href="#">Back</a>
+                <a href="/login">Back</a>
             </div>
 
             <div class="form-container">
@@ -136,25 +143,7 @@
 
 
                     <div class="col-lg-12">
-                        <button class="btn btn-outline-primary" type="submit">Submit form </button>
-                    </div>
-
-
-                    <div class="col-md-6 form-floating mb-4">
-                        <input type="email" class="form-control" id="floatingemail" placeholder="Email Address" required>
-                        <label for="floatingemail">Email Address</label>
-
-                    </div>
-
-                    <div class="col-md-6 form-floating mb-4">
-                        <input type="password" class="form-control" id="floatingPassword" placeholder="Password" required>
-                        <label for="floatingPassword">Password</label>
-                    </div>
-
-                    
-
-                    <div class="col-12">
-                        <button class="btn btn-primary" type="submit">Submit form</button>
+                        <button class="btn btn-outline-primary" type="submit">Submit form</button>
                     </div>
 
                 </form>
