+++ admin-system/src/main/resources/templates/PatientWebPages/index.html
@@ -8,6 +8,6 @@
 </head>
     <body>
         <h1>Welcome to the landing page!</h1>
-        <a href="">Patient Login Page</a>
+        <a href="/login">Patient Login Page</a>
     </body>
 </html>
