+++ admin-system/src/main/resources/templates/admin/Login.html
@@ -5,11 +5,12 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Gillego-Flores Dental Clinic</title>
+
     <link rel="stylesheet" href="/css/admin/LoginForm.css">
     <link rel="shortcut icon" href="favico n.ico" type="image/x-icon">
 
-    <link rel="icon" type="image/png" sizes="32x32" href="/assets/new-resized-logo.png">
-    <link rel="shortcut icon" type="image/x-icon" sizes="16x16" href="/assets/new-resized-logo.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/assets/administrator/new-resized-logo.png">
+    <link rel="shortcut icon" type="image/x-icon" sizes="16x16" href="/assets/administrator/new-resized-logo.png">
     <link rel="manifest" href="/site.webmanifest">
     <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5">
     <meta name="msapplication-TileColor" content="#da532c">
@@ -30,10 +31,10 @@
                         <input type="password" name="password" placeholder="Password">
                         <input type="submit" value="Login">
                     </form>
-                    <p>Register Here <a href="/system/admin/registration">Register</a></p>
+                    <p>Register Here <a href="/admin/registration">Register</a></p>
                 </div>
                 <div class="container-img">
-                    <img src="/assets/person-managing-it-settings.png" alt="Main-image">
+                    <img src="/assets/administrator/person-managing-it-settings.png" alt="Main-image">
                 </div>
             </div>
         </div>
