+++ admin-system/src/main/resources/templates/admin/Registration.html
@@ -7,12 +7,13 @@
     <title>Gillego-Flores Dental Clinic</title>
     <link rel="stylesheet" href="/css/admin/RegistrationForm.css">
     
-    <link rel="icon" type="image/png" sizes="32x32" href="/assets/new-resized-logo.png">
-    <link rel="shortcut icon" type="image/x-icon" sizes="16x16" href="/assets/new-resized-logo.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/assets/administrator/new-resized-logo.png">
+    <link rel="shortcut icon" type="image/x-icon" sizes="16x16" href="/assets/administrator/new-resized-logo.png">
     <link rel="manifest" href="/site.webmanifest">
     <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5">
     <meta name="msapplication-TileColor" content="#da532c">
     <meta name="theme-color" content="#ffffff">
+
 </head>
     <body>
         <div class="container">
@@ -98,10 +99,10 @@
                     </div>
                     <div class="buttons"> 
                         <div class="back-button">
-                            <input type="button" value="Back" onclick="location.href='/system/admin/login' ">
+                            <input type="button" value="Back" onclick="location.href='/admin/login' ">
                         </div>
                         <div class="submit-button">
-                            <input type="submit" value="Submit" onclick="location.href='/system/admin/login'">
+                            <input type="submit" value="Submit" onclick="location.href='/admin/login'">
                         </div>
                     </div>
                 </form>
