+++ admin-system/src/main/resources/templates/dashboard/Dashboard.html
@@ -10,20 +10,22 @@
         <link rel="stylesheet" href="/css/admin/dashboard.css">
    
 
-    <!-- with out localhost -->
+    <!-- without localhost -->
     <!-- <link rel="stylesheet" href="/admin-system/src/main/resources/static/css/dashboard.css"> -->
    
     <!-- fontawesome link -->
+
     <!-- <link href="http://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.3.0/css/font-awesome.css" 
     rel="stylesheet"  type='text/css'>
     <script src="https://kit.fontawesome.com/139b947450.js" crossorigin="anonymous"></script> -->
     
-    <link rel="icon" type="image/png" sizes="32x32" href="/assets/new-resized-logo.png">
-    <link rel="shortcut icon" type="image/x-icon" sizes="16x16" href="/assets/new-resized-logo.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/assets/administrator/new-resized-logo.png">
+    <link rel="shortcut icon" type="image/x-icon" sizes="16x16" href="/assets/administrator/new-resized-logo.png">
     <link rel="manifest" href="/site.webmanifest">
     <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5">
     <meta name="msapplication-TileColor" content="#da532c">
     <meta name="theme-color" content="#ffffff">
+
     <script type="text/javascript" src=\javascript\dashboard.js></script>
 
 </head>
@@ -33,7 +35,7 @@
         <nav class="main-container-side-navbar">
             
             <div class="admin-logo">
-                <img src="/assets/user-image.png" alt="user-image">
+                <img src="/assets/administrator/user-image.png" alt="user-image">
                 
                 <div class="admin-logo-title">
                     <h1>ADMINISTRATOR</h1>
@@ -43,22 +45,22 @@
 
             <ul>
                 <li>
-                    <img src="/assets/dashboard.png" alt="dashboa-image">
+                    <img src="/assets/administrator/dashboard.png" alt="dashboard-image">
                     <a href="#">Dashboard</a>
                 </li>
                 
                 <li>
-                    <img src="/assets/Clients.png" alt="client-image">
+                    <img src="/assets/administrator/Clients.png" alt="client-image">
                     <a href="#">Clients</a>
                 </li>
 
                 <li>
-                    <img src="/assets/cancelled.png" alt="cancelled-request-image">
+                    <img src="/assets/administrator/cancelled.png" alt="cancelled-request-image">
                     <a href="#">Cancelled Client Request</a>
                 </li>
 
                 <li>
-                    <img src="/assets/checklist.png" alt="client-request-image">
+                    <img src="/assets/administrator/checklist.png" alt="client-request-image">
                     <a href="#">Client Request Service</a>
                 </li>
 
@@ -71,7 +73,7 @@
 
         <main class="main">
                 <div class="main-heading">
-                    <img src="\assets\background-remove-logo.png" alt="logo">
+                    <img src="/assets/administrator/background-remove-logo.png" alt="logo">
                     <div class="main-heading-title">
                         <h1>Gillego-Flores Dental Clinic</h1>
                         <p>Dashboard</p>
@@ -82,19 +84,19 @@
                     <div class="main-container-list">
                         
                         <div class="main-container-list-client">
-                            <img src="/assets/Clients.png" alt="Client">
+                            <img src="/assets/administrator/Clients.png" alt="Client">
                             <h2>List of CLients</h2>
                             <p><span>0</span></p>
                         </div>
                                                  
                         <div class="main-container-list-pending">
-                            <img src="/assets/checklist.png" alt="Pending">
+                            <img src="/assets/administrator/checklist.png" alt="Pending">
                             <h2>Pending Request</h2>
                             <p><span>0</span></p>
                         </div>
                         
                         <div class="main-container-list-cancelled"> 
-                            <img src="/assets/cancelled.png" alt="Cancelled">
+                            <img src="/assets/administrator/cancelled.png" alt="Cancelled">
                             <h2>Cancelled Request</h2>
                             <p><span>0</span></p>
                         </div>
